+++ atd_knack_services_sm_contractors-deployment.yaml
@@ -1,0 +1,69 @@
+###
+### A complete description of a Prefect Deployment for flow 'Knack Services: Signs Markings Contractor Work Orders'
+###
+name: 'Knack Services: SM Contractor Work Orders'
+description: null
+version: 2ab508ab01c6ae19fd371f9e2672706a
+# The work queue that will handle this deployment's runs
+work_queue_name: default
+work_pool_name: atd-data-03
+tags: []
+parameters:
+  block: atd-knack-services-sm-contractor-work-orders
+  commands:
+  - atd-knack-services/services/records_to_postgrest.py -a signs-markings -c view_3628
+  - atd-knack-services/services/records_to_agol.py -a signs-markings -c view_3628
+  - atd-knack-services/services/records_to_socrata.py -a signs-markings -c view_3628
+  - atd-knack-services/services/agol_build_markings_segment_geometries.py -l markings_contractor_work_orders
+schedule:
+  cron: 5 7 * * *
+  timezone: null
+  day_or: true
+is_schedule_active: true
+infra_overrides: {}
+
+###
+### DO NOT EDIT BELOW THIS LINE
+###
+flow_name: 'Knack Services: Signs Markings Contractor Work Orders'
+manifest_path: null
+infrastructure:
+  type: process
+  env: {}
+  labels: {}
+  name: null
+  command: null
+  stream_output: true
+  working_dir: null
+  _block_document_id: 94db8c96-2455-453a-b143-845ffa98ea2a
+  _block_document_name: anonymous-3dd57c23-aaf4-4ebb-9353-ec6d51f79176
+  _is_anonymous: true
+  block_type_slug: process
+  _block_type_slug: process
+storage:
+  repository: https://github.com/cityofaustin/atd-prefect
+  reference: ch-prefect2-atd-knack-services
+  access_token: null
+  include_git_objects: true
+  _block_document_id: a404f1cd-6b78-49ca-a0ca-ada4d56f3482
+  _block_document_name: knack-services-wip
+  _is_anonymous: false
+  block_type_slug: github
+  _block_type_slug: github
+path: ''
+entrypoint: flows/knack/atd_knack_markings_contractor_work_orders.py:atd_knack_services_sm_contractors
+parameter_openapi_schema:
+  title: Parameters
+  type: object
+  properties:
+    commands:
+      title: commands
+      position: 0
+    block:
+      title: block
+      position: 1
+  required:
+  - commands
+  - block
+  definitions: null
+timestamp: '2023-02-22T17:19:40.286868+00:00'
